@@ -1,12 +1,22 @@
-module github.com/Pacific73/gorm-cache
+module github.com/swordkee/gorm-cache
 
-go 1.15
+go 1.18
 
 require (
-	github.com/go-redis/redis/v8 v8.11.5
-	github.com/karlseguin/ccache/v2 v2.0.8
-	github.com/onsi/gomega v1.19.0 // indirect
+	github.com/karlseguin/ccache/v3 v3.0.5
+	github.com/redis/go-redis/v9 v9.2.1
 	github.com/smartystreets/goconvey v1.7.2
-	gorm.io/driver/mysql v1.3.3
-	gorm.io/gorm v1.23.4
+	gorm.io/driver/mysql v1.4.5
+	gorm.io/gorm v1.24.6
+)
+
+require (
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/go-sql-driver/mysql v1.7.0 // indirect
+	github.com/gopherjs/gopherjs v0.0.0-20181017120253-0766667cb4d1 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.5 // indirect
+	github.com/jtolds/gls v4.20.0+incompatible // indirect
+	github.com/smartystreets/assertions v1.2.0 // indirect
 )
