@@ -1,9 +1,8 @@
 package config
 
 import (
+	"github.com/redis/go-redis/v9"
 	"sync"
-
-	"github.com/go-redis/redis/v8"
 )
 
 type RedisConfigMode int
@@ -16,16 +15,16 @@ const (
 type RedisConfig struct {
 	Mode RedisConfigMode
 
-	Options *redis.Options
-	Client  *redis.Client
+	Options *redis.UniversalOptions
+	Client  redis.UniversalClient
 
 	once sync.Once
 }
 
-func (c *RedisConfig) InitClient() *redis.Client {
+func (c *RedisConfig) InitClient() redis.UniversalClient {
 	c.once.Do(func() {
 		if c.Mode == RedisConfigModeOptions {
-			c.Client = redis.NewClient(c.Options)
+			c.Client = redis.NewUniversalClient(c.Options)
 		}
 	})
 	return c.Client
