@@ -7,10 +7,10 @@ import (
 
 	"gorm.io/gorm/logger"
 
-	"github.com/Pacific73/gorm-cache/cache"
+	"github.com/swordkee/gorm-cache/cache"
 
-	"github.com/Pacific73/gorm-cache/config"
-	"github.com/go-redis/redis/v8"
+	"github.com/redis/go-redis/v9"
+	"github.com/swordkee/gorm-cache/config"
 	"gorm.io/driver/mysql"
 	"gorm.io/gorm"
 )
@@ -87,7 +87,7 @@ func TestMain(m *testing.M) {
 		os.Exit(-1)
 	}
 
-	redisClient := redis.NewClient(&redis.Options{Addr: redisIp + ":" + redisPort})
+	redisClient := redis.NewUniversalClient(&redis.UniversalOptions{Addrs: []string{redisIp + ":" + redisPort}})
 
 	searchCache, err = cache.NewGorm2Cache(&config.CacheConfig{
 		CacheLevel:           config.CacheLevelOnlySearch,
@@ -134,9 +134,9 @@ func TestMain(m *testing.M) {
 		os.Exit(-1)
 	}
 
-	primaryDB.Use(primaryCache)
-	searchDB.Use(searchCache)
-	allDB.Use(allCache)
+	_ = primaryDB.Use(primaryCache)
+	_ = searchDB.Use(searchCache)
+	_ = allDB.Use(allCache)
 	// primaryCache.AttachToDB(primaryDB)
 	// searchCache.AttachToDB(searchDB)
 	// allCache.AttachToDB(allDB)
